--- src/main/java/gift/Message.java
@@ -1,7 +1,0 @@
-package gift;
-
-public class Message {
-    public static final String ADD_SUCCESS_MSG = "상품 추가 성공";
-    public static final String UPDATE_SUCCESS_MSG = "상품 수정 성공";
-    public static final String DELETE_SUCCESS_MSG = "상품 삭제 성공";
-}
