--- src/main/java/gift/ProductController.java
@@ -1,35 +1,0 @@
-package gift;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/api/products")
-public class ProductController {
-    private final ProductService productService;
-    public ProductController(ProductService productService) {
-        this.productService = productService;
-    }
-
-    @GetMapping("/{productId}")
-    public Product retrieveProduct(@PathVariable("productId") Long productId) {
-        return productService.getOneProduct(productId);
-    }
-
-    @PostMapping
-    public ResponseEntity<String> addProduct(@RequestBody Product product) {
-        return ResponseEntity.ok(productService.addProduct(product));
-    }
-
-    @PutMapping("/{productId}")
-    public ResponseEntity<String> updateProduct(@PathVariable("productId") Long productId, @RequestBody Product product) {
-        return ResponseEntity.ok(productService.updateProduct(productId, product));
-    }
-
-    @DeleteMapping("/{productId}")
-    public ResponseEntity<String> deleteProduct(@PathVariable("productId") Long productId) {
-        return ResponseEntity.ok(productService.deleteProduct(productId));
-    }
-
-}
-
