+++ src/main/java/gift/controller/ProductController.java
@@ -1,7 +1,7 @@
 package gift.controller;
 
-import gift.Product;
-import gift.ProductService;
+import gift.domain.Product;
+import gift.service.ProductService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
