+++ src/main/java/gift/controller/ProductManageController.java
@@ -1,7 +1,7 @@
 package gift.controller;
 
-import gift.Product;
-import gift.ProductService;
+import gift.domain.Product;
+import gift.service.ProductService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
