+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package gift;
+package gift.repository;
 
+import gift.domain.Product;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
